freebusy: use size_t for RecurrenceState::ParseBlob

+++ swig/RecurrenceState.swig.h
@@ -99,7 +99,7 @@ public:
 
 class RecurrenceState {
 public:
-	HRESULT ParseBlob(const char *lpData, unsigned int ulLen, ULONG ulFlags);
+	HRESULT ParseBlob(const char *lpData, size_t ulLen, ULONG ulFlags);
 	HRESULT GetBlob(char **lpData, size_t *lpulLen, void *base = NULL);
 
 	/* not inlined for swig */
